# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RESTful API service for managing users, websites, and pages with session-based authentication.",
        "title": "XeoDocs Dash API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "XeoDocs Team",
            "url": "http://www.xeodocs.com",
            "email": "support@xeodocs.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/websites": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get list of all websites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websites"
                ],
                "summary": "Get all websites",
                "responses": {
                    "200": {
                        "description": "List of websites",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Website"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new website",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websites"
                ],
                "summary": "Create new website",
                "parameters": [
                    {
                        "description": "Website creation data",
                        "name": "website",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateWebsiteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Website created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/models.Website"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/websites/slug/{slug}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a specific website by its slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websites"
                ],
                "summary": "Get website by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Website slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Website details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/models.Website"
                            }
                        }
                    },
                    "404": {
                        "description": "Website not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/websites/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a specific website by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websites"
                ],
                "summary": "Get website by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Website ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Website details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/models.Website"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid website ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Website not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update website information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websites"
                ],
                "summary": "Update website",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Website ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Website update data",
                        "name": "website",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateWebsiteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Website updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/models.Website"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Website not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a website",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websites"
                ],
                "summary": "Delete website",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Website ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Website deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid website ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Website not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateWebsiteRequest": {
            "type": "object",
            "required": [
                "config",
                "description",
                "domain",
                "git_api_token",
                "git_repo_branch",
                "git_repo_name",
                "git_repo_owner",
                "language_code",
                "name",
                "slogan",
                "slug"
            ],
            "properties": {
                "config": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "git_api_token": {
                    "type": "string"
                },
                "git_repo_branch": {
                    "type": "string"
                },
                "git_repo_name": {
                    "type": "string"
                },
                "git_repo_owner": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slogan": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateWebsiteRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "git_api_token": {
                    "type": "string"
                },
                "git_repo_branch": {
                    "type": "string"
                },
                "git_repo_name": {
                    "type": "string"
                },
                "git_repo_owner": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slogan": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "models.Website": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "git_repo_branch": {
                    "type": "string"
                },
                "git_repo_name": {
                    "type": "string"
                },
                "git_repo_owner": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "language_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slogan": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer token for authentication. Format: 'Bearer {token}'",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}